GH-683: additional fix to renaming
