+++ node/src/blockchain/errors/blockchain_loggable_error/masq_error.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2025, MASQ (https://masq.ai) and/or its affiliates. All rights reserved.
 
 use crate::blockchain::errors::blockchain_db_error::BlockchainDbError;
-use crate::blockchain::errors::blockchain_error::BlockchainLoggableError;
+use crate::blockchain::errors::blockchain_loggable_error::BlockchainLoggableError;
 use crate::blockchain::errors::custom_common_methods::CustomCommonMethods;
 use std::fmt::{Debug, Display, Formatter};
 
@@ -41,12 +41,8 @@ impl Display for MASQError {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::blockchain::errors::blockchain_error::app_rpc_web3_error::{
-        AppRpcWeb3Error, LocalError, RemoteError,
-    };
-    use crate::blockchain::errors::blockchain_error::BlockchainLoggableError;
+    use crate::blockchain::errors::blockchain_loggable_error::BlockchainLoggableError;
     use crate::blockchain::errors::test_utils::test_clone_impl_for_blockchain_error;
-    use std::fmt::format;
 
     #[test]
     fn clone_works_for_blockchain_error_wrapping_masq_error() {
