+++ node/src/blockchain/errors/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2025, MASQ (https://masq.ai) and/or its affiliates. All rights reserved.
 
 pub mod blockchain_db_error;
-pub mod blockchain_error;
+pub mod blockchain_loggable_error;
 mod custom_common_methods;
 mod test_utils;
 pub mod validation_status;
